added unfurl preview card for rewards, added share reward button for each reward
