+++ lib/player-lists/public-list-og-card.ts
@@ -1,6 +1,6 @@
 import type { PublicCustomListWithLocations } from '@/lib/db/player-custom-lists';
 import { resolvePublicListShareLookup } from '@/lib/db/player-custom-lists';
-import { supabaseRenderImageUrl } from '@/lib/metadata/request-base';
+import { ogCardPhotoUrl } from '@/lib/og/satori-card-assets';
 import { isReservedUsername, normalizeUsername } from '@/lib/username';
 
 export type PublicListOgCard = {
@@ -50,25 +50,11 @@ export function buildPublicListOgCard(
   };
 }
 
-/**
- * Photo URL Satori can decode (PNG/JPEG/GIF). Supabase WebP is converted via
- * image transforms; other WebP sources are skipped so the card can fall back.
- */
+/** Photo URL Satori can decode, or null when the card should fall back. */
 export function publicListOgSatoriPhotoUrl(
   photoUrl: string | null
 ): string | null {
-  const trimmed = photoUrl?.trim();
-  if (!trimmed) return null;
-
-  try {
-    const rendered = supabaseRenderImageUrl(trimmed);
-    if (rendered) return rendered;
-  } catch {
-    return null;
-  }
-
-  if (/\.webp(?:$|\?)/i.test(trimmed)) return null;
-  return trimmed;
+  return ogCardPhotoUrl(photoUrl);
 }
 
 /**
